+++ magentoerpconnect/product.py
@@ -27,7 +27,7 @@ import xmlrpclib
 import sys
 from openerp.osv import orm, fields
 from openerp.tools.translate import _
-from openerp.addons.connector.queue.job import job
+from openerp.addons.connector.queue.job import job, related_action
 from openerp.addons.connector.event import on_record_write
 from openerp.addons.connector.connector import ConnectorUnit
 from openerp.addons.connector.unit.synchronizer import (ImportSynchronizer,
@@ -50,6 +50,7 @@ from .unit.import_synchronizer import (DelayedBatchImport,
                                        )
 from .connector import get_environment
 from .backend import magento
+from .related_action import unwrap_binding
 
 _logger = logging.getLogger(__name__)
 
@@ -177,6 +178,7 @@ class product_product(orm.Model):
 class ProductProductAdapter(GenericAdapter):
     _model_name = 'magento.product.product'
     _magento_model = 'catalog_product'
+    _admin_path = '/{model}/edit/id/{id}'
 
     def _call(self, method, arguments):
         try:
@@ -460,7 +462,8 @@ class ProductImport(MagentoImportSynchronizer):
         """ Hook called at the end of the import """
         translation_importer = self.get_connector_unit_for_model(
             TranslationImporter, self.model._name)
-        translation_importer.run(self.magento_id, binding_id)
+        translation_importer.run(self.magento_id, binding_id,
+                                 mapper_class=ProductImportMapper)
         image_importer = self.get_connector_unit_for_model(
             CatalogImageImporter, self.model._name)
         image_importer.run(self.magento_id, binding_id)
@@ -468,6 +471,18 @@ class ProductImport(MagentoImportSynchronizer):
         if self.magento_record['type_id'] == 'bundle':
             bundle_importer = self.get_connector_unit_for_model(
                 BundleImporter, self.model._name)
+
+
+@magento
+class IsActiveProductImportMapper(ImportMapper):
+    _model_name = 'magento.product.product'
+
+    @mapping
+    def is_active(self, record):
+        """Check if the product is active in Magento
+        and set active flag in OpenERP
+        status == 1 in Magento means active"""
+        return {'active': (record.get('status') == '1')}
 
 
 @magento
@@ -626,6 +641,7 @@ def magento_product_modified(session, model_name, record_id, vals):
 
 
 @job
+@related_action(action=unwrap_binding)
 def export_product_inventory(session, model_name, record_id, fields=None):
     """ Export the inventory configuration and quantity of a product. """
     product = session.browse(model_name, record_id)