+++ magentoerpconnect/sale.py
@@ -22,8 +22,9 @@
 import logging
 import xmlrpclib
 from datetime import datetime, timedelta
-from openerp.osv import fields, orm
 import openerp.addons.decimal_precision as dp
+from openerp.osv import fields, orm
+from openerp.tools.translate import _
 from openerp.addons.connector.connector import ConnectorUnit
 from openerp.addons.connector.exception import (NothingToDoJob,
                                                 FailedJobError,
@@ -211,6 +212,7 @@ class sale_order_line(orm.Model):
 class SaleOrderAdapter(GenericAdapter):
     _model_name = 'magento.sale.order'
     _magento_model = 'sales_order'
+    _admin_path = '{model}/view/order_id/{id}'
 
     def _call(self, method, arguments):
         try:
@@ -371,6 +373,21 @@ class SaleOrderImport(MagentoImportSynchronizer):
        if self._mapper is None:
            self._mapper = self.environment.get_connector_unit(SaleOrderImportMapper)
        return self._mapper
+
+    def _must_skip(self):
+        """ Hook called right after we read the data from the backend.
+
+        If the method returns a message giving a reason for the
+        skipping, the import will be interrupted and the message
+        recorded in the job (if the import is called directly by the
+        job, not by dependencies).
+
+        If it returns None, the import will continue normally.
+
+        :returns: None | str | unicode
+        """
+        if self.binder.to_openerp(self.magento_id):
+            return _('Already imported')
 
     def _clean_magento_items(self, resource):
         """
@@ -545,7 +562,14 @@ class SaleOrderImport(MagentoImportSynchronizer):
                 partner = sess.read('magento.res.partner',
                                     partner_ids[0],
                                     ['magento_id'])
-                record['customer_id'] = partner['magento_id']
+                # If there are multiple orders with (customer_id is null) and (customer_is_guest = 0)
+                #  which share the same customer_email, then we may get a magento_id
+                #  that is a marker 'guestorder:...' for a guest order (which is set below).
+                #  This causes a problem with "importer.run..." below where the id is cast to int.
+                if str(partner['magento_id']).startswith('guestorder:'):
+                    is_guest_order = True
+                else:
+                    record['customer_id'] = partner['magento_id']
 
             # no partner matching, it means that we have to consider it
             # as a guest order